# doc/test.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "test.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/base/coding": {
      "post": {
        "operationId": "MakeCoding",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/echopbTestResp"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/echopbMakeCodingRequest"
            }
          }
        ],
        "tags": [
          "Test"
        ]
      }
    },
    "/v1/test": {
      "get": {
        "summary": "测试方法",
        "operationId": "GetTestMsg",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/echopbTestResp"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "value",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Test"
        ]
      }
    }
  },
  "definitions": {
    "echopbMakeCodingRequest": {
      "type": "object",
      "properties": {
        "serverName": {
          "type": "string"
        },
        "moduleName": {
          "type": "string"
        },
        "tableName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "databaseName": {
          "type": "string"
        }
      }
    },
    "echopbTestDetail": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
    "echopbTestResp": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "details": {
          "$ref": "#/definitions/echopbTestDetail"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}